microservices/inventory/pyproject.toml
[project]
name = "inventory"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "fastapi==0.128.0",
    "python-dotenv>=1.2.1",
    "redis-om==0.3.5",
]
